pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "icav2-cli-plugins"
version = "2.27.0.dev20240710144741"
authors = [
    { name = "Alexis Lucattini", email = "alexis.lucattini@umccr.org" },
]
description = "icav2-cli-plugins"
readme = "Readme.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "beautifulsoup4 >= 4.11.1, < 5",
    "cwl_utils >= 0.32, < 1",
    "docopt >= 0.6.2, < 1",
    "libica >= 2.4.0, < 3",
    "mdutils >= 1.4.0, < 2",
    "pandas >= 2.1.3, < 3",
    "PyJWT >= 2.6.0, < 3",
    "ruamel.yaml >= 0.18, < 0.19",
    "requests >= 2.31.0, < 3",
    "ruamel.base == 1.0.0",
    "tabulate >= 0.9.0, < 1",
    "verboselogs >= 1.7, < 2",
    "websocket_client >= 1.4.2, < 2",
    "cwltool >= 3.1.20231016170136, < 4",
    "pandoc >= 2.3, < 3",
    "humanfriendly >= 10.0, < 11",
    "matplotlib >= 3.7.1, < 4",
    "invoke >= 2.0.0, < 3",
    "fabric >= 3.0.0, < 4",
    "deepdiff >= 6.7.1, < 7",
    "wrapica >= 2.27.1, <2.28.0",
    "nf-core >= 2.14",
]

[project.urls]
Homepage = "https://github.com/umccr/icav2-cli-plugins"
"Bug Tracker" = "https://github.com/umccr/icav2-cli-plugins/issues"

[project.scripts]
"icav2-cli-plugins.py" = "icav2_cli_plugins.utils.cli:main"

[project.optional-dependencies]
pandoc = [
    "pypandoc==1.10",
    "pypandoc_binary==1.10",
]
toml = [
    "tomli_w >= 1.0.0, < 2",
]
docs = [
    "sphinx >= 7.2.6, < 8",
    "sphinx-rtd-theme >= 2.0.0, < 3",
    "sphinx_autodoc_typehints",
    "toml-to-requirements",
]
